add LFV_ROOT_FOLDER to disallow user from going past that
