--- src/lib/server-utils/base-directory.ts
@@ -1,15 +1,0 @@
-let baseDirectory = '';
-
-if (process.env.LFV_DEFAULT_FOLDER) {
-	baseDirectory = process.env.LFV_DEFAULT_FOLDER;
-} else {
-	throw new Error('No base directory provided.');
-}
-
-export function getBaseDirectory() {
-	return baseDirectory;
-}
-
-export function setBaseDirectory(dir: string) {
-	baseDirectory = dir;
-}
