+++ src/routes/serve/[...file]/+server.ts
@@ -2,7 +2,7 @@ import path from 'path';
 
 import { readFile } from '../../../lib/server-utils/mem-fs';
 import { getMimeType } from '../../../lib/server-utils/get-mime-types';
-import { getBaseDirectory } from '../../../lib/server-utils/base-directory';
+import { getBaseDirectory } from '../../../lib/server-utils/directory-variables';
 
 export const GET = async function ({ params }: { params: any }) {
 	const filePath = path.join(getBaseDirectory(), params.file);
